requirements.txt
requests>=2.31.0
GitPython>=3.1.40
PyYAML>=6.0
click>=8.1.7
rich>=13.7.0
python-dateutil>=2.8.2
